Move TemplateFunction type
