+++ apps/pie-storybook/stories/pie-toggle-switch.stories.ts
@@ -3,8 +3,7 @@ import { PieToggleSwitch, ToggleSwitchProps, labelPlacements } from '@justeattak
 import { IconCheck } from '@justeattakeaway/pie-icons-webc';
 import { StoryMeta } from '../types';
 import { i18nArgTypes } from '../args/commonArgsTypes';
-import { TemplateFunction } from '../types/StoryOptions';
-import { createStory } from '../utilities';
+import { createStory, type TemplateFunction } from '../utilities';
 
 // This prevents storybook from tree shaking the components
 // eslint-disable-next-line @typescript-eslint/no-unused-vars
