+++ apps/pie-storybook/utilities/index.ts
@@ -1,4 +1,7 @@
-import { TemplateFunction, StoryOptions } from '../types/StoryOptions';
+import { TemplateResult } from 'lit';
+import { StoryOptions } from '../types/StoryOptions';
+
+export type TemplateFunction<T> = (props: T) => TemplateResult;
 
 export const createStory = <T>(templateFunc: TemplateFunction<T>, defaultArgs: T) => (
     propOverrides?: Partial<T>,
